Move to pyproject.toml
* Add pyproject.toml configuration
* Update main entrypoint to work with pyproject.toml
* Remove requirements.txt and setup.py files
* Update Dockerfile to install non-editable version of the package
* Update self package imports to use common root of stac.
