+++ .dockerignore
@@ -1,0 +1,5 @@
+__pycache__/
+.pytest_cache/
+.mypy_cache/
+*.pyc
+stac_fastapi_k8s/
