+++ pyproject.toml
@@ -1,0 +1,102 @@
+# --------------------------------------------------------------------------------------------
+# Copyright (c) Microsoft Corporation. All rights reserved.
+# Licensed under the MIT License. See License.txt in the project root for license information.
+# --------------------------------------------------------------------------------------------
+[project]
+name = "azure-stac-cli"
+version = "0.1.0"
+description = "STAC Reference Architecture"
+readme = "README.md"
+requires-python = ">= 3.8"
+classifiers = [
+  "Framework :: FastAPI",
+  "Operating System :: OS Independent",
+  "Programming Language :: Python :: 3 :: Only",
+  "Programming Language :: Python :: 3.8",
+  "Programming Language :: Python :: 3.9",
+  "Programming Language :: Python :: 3.10",
+  "License :: OSI Approved :: MIT License",
+  "Topic :: Scientific/Engineering :: GIS",
+]
+dependencies = [
+  "knack",
+  "opencensus-ext-azure",
+  "azure-servicebus == 7.8.1",
+  "azure-storage-blob == 12.14.1",
+  "pystac == 1.6.1",
+  "python_dateutil == 2.8.2",
+  "python_dotenv == 0.21.0",
+  "rasterio == 1.3.4",
+  "stactools == 0.4.2",
+  "psycopg[binary] == 3.1.7",
+  "pypgstac[psycopg] == 0.6.11",
+  "charset-normalizer <3.0, >=2.0"
+]
+
+[project.urls]
+repository = "https://github.com/Azure/Azure-Orbital-STAC"
+
+[project.optional-dependencies]
+dev = [
+  "black >= 22.12.0",
+  "isort >= 5.11.0",
+  "mypy >= 0.991",
+  "flake8 >= 6.0.0",
+  "flake8-pyproject >= 1.2.2",  # for flake8 to read pyproject.toml
+]
+
+[project.scripts]
+stac = "azure_stac.cli.__main__:main"
+
+[build-system]
+requires = ["setuptools >= 40.9.0", "wheel"]
+build-backend = "setuptools.build_meta"
+
+[tool.setuptools.packages.find]
+where = ["src"]
+include = ["azure_stac.*"]
+
+[tool.flake8]
+max-line-length = 96
+per-file-ignores = [
+  # Ignore unused imports in __init__.py files
+  '__init__.py:F401'
+]
+exclude = ['.git', '__pycache__', '.mypy_cache', 'build', 'dist']
+
+[tool.black]
+line-length = 96
+target-version = ['py37']
+include = '\.pyi?$'
+exclude = '''
+(
+  /(
+      \.eggs         # exclude a few common directories in the
+    | \.git          # root of the project
+    | \.hg
+    | \.mypy_cache
+    | \.tox
+    | \.venv
+    | _build
+    | buck-out
+    | build
+    | dist
+  )/
+)
+'''
+
+[tool.mypy]
+files = ["src"]
+disallow_untyped_defs = true
+ignore_missing_imports = true
+follow_imports = "skip"
+strict_optional = true
+
+[tool.isort]
+profile = "black"
+skip = [
+    '.eggs', '.git', '.hg', '.mypy_cache', '.nox', '.pants.d', '.tox',
+    '.venv', '_build', 'buck-out', 'build', 'dist', 'node_modules', 'venv',
+]
+skip_gitignore = true
+line_length = 96
