--- src/azure-stac-cli/stac/metrics/pod_metrics.py
@@ -1,22 +1,0 @@
-# --------------------------------------------------------------------------------------------
-# Copyright (c) Microsoft Corporation. All rights reserved.
-# Licensed under the MIT License. See License.txt in the project root for license information.
-# --------------------------------------------------------------------------------------------
-
-import os
-
-from stac.core.metrics import Metrics
-
-class PodMetrics(Metrics):
-    
-    def __init__(self, *args, **kwargs):
-        
-        pass
-    
-    def register_metrics(self):
-        
-        pass
-    
-    def send_metrics(self):
-        
-        pass
