--- src/azure-stac-cli/stac/stac
@@ -1,24 +1,0 @@
-#!/usr/bin/env python
-
-# --------------------------------------------------------------------------------------------
-# Copyright (c) Microsoft Corporation. All rights reserved.
-# Licensed under the MIT License. See License.txt in the project root for license information.
-# --------------------------------------------------------------------------------------------
-
-import sys
-import os
-
-dir = os.path.join(os.path.dirname(os.path.realpath(__file__)), 'src')
-
-if os.environ.get('PYTHONPATH') is None:
-    os.environ['PYTHONPATH'] = dir
-else:
-    os.environ['PYTHONPATH'] = os.pathsep.join([
-        dir,
-        os.environ.get('PYTHONPATH'),
-    ])
-
-if os.environ.get('STAC_INSTALLER') is None:
-    os.environ['STAC_INSTALLER'] = 'PIP'
-
-os.execl(sys.executable, sys.executable, '-m', 'cli', *sys.argv[1:])
