+++ src/azure_stac/commands/__command.py
