+++ src/azure_stac/common/__init__.py
