+++ src/azure_stac/common/__utilities.py
