+++ src/azure_stac/processors/core/__init__.py
