+++ src/azure_stac/processors/naip/__constants.py
